+++ src/IECoreDelight/Display.cpp
@@ -114,9 +114,23 @@ PtDspyError imageOpen( PtDspyImageHandle *image, const char *driverName, const c
 			const string layerName = *(const char **)(parameters[p].value);
 			if( !layerName.empty() )
 			{
-				for( auto &channel : channels )
+				if( channels.size() == 1 )
 				{
-					channel = layerName + "." + channel;
+					// I'm not sure what the semantics of 3Delight's `layername`
+					// actually are, but this gets the naming matching Arnold
+					// for our all-important OutputBuffer outputs used in the
+					// Viewer.
+					/// \todo We're overdue a reckoning were we define our own
+					/// standard semantics for all the little details of outputs,
+					/// and implement them to match across all renderers.
+					channels[0] = layerName;
+				}
+				else
+				{
+					for( auto &channel : channels )
+					{
+						channel = layerName + "." + channel;
+					}
 				}
 			}
 		}
@@ -300,7 +314,29 @@ PtDspyError imageData( PtDspyImageHandle image, int xMin, int xMaxPlusOne, int y
 	{
 		try
 		{
-			dd->imageData( box, (const float *)data, bufferSize );
+			if( dd->channelNames().size() == 1 && dd->channelNames()[0] == "id" )
+			{
+				// ID output for doing interactive selection (see `GafferSceneUI::OutputBuffer`).
+				// We want to have declared this output with an integer `scalarformat` to match the
+				// integer `cortexId` attribute created by `ObjectInterface::assignID()`, but we
+				// can't because 3Delight thinks it's quantizing a float into an integer and throws
+				// away everything outside `0-1` (see DelightOutput). So, we instead get floats
+				// from 3Delight and convert to integers here.
+				vector<uint32_t> intData( bufferSize );
+				for( int i = 0; i < bufferSize; ++i )
+				{
+					intData[i] = ((const float *)data)[i];
+				}
+				// Then, in a second hack that applies to all renderers/drivers,
+				// we have to pretend that the integers are floats to get them
+				// across the `IECoreImage::DisplayDriver` boundary, which only
+				// supports floats.
+				dd->imageData( box, (const float *)intData.data(), bufferSize );
+			}
+			else
+			{
+				dd->imageData( box, (const float *)data, bufferSize );
+			}
 		}
 		catch( std::exception &e )
 		{